Refactored ls.rs to fix cargo fmt
`cargo fmt` have trouble formatting `clap::Arg`.
This commit simply splits some stuff up so that the implementation details
can be formatted.
